requirements.txt
deepface>=0.0.86
opencv-python>=4.9.0
streamlit>=1.31.0
numpy>=1.26.0
tf-keras>=2.15.0
tensorflow>=2.15.0
pillow>=10.2.0
matplotlib>=3.8.0
